recommender/requirements.txt
Flask==2.2.3
Werkzeug==2.3.7
requests
redis
psycopg2-binary